refactor(server): cache marketplace catalog via CoreEntityCacheProvider
Move the marketplace catalog cache onto the CoreEntityCache infrastructure
through a MarketplaceCatalogCacheProviderService. The whole catalog is
cached as a single Record<universalIdentifier, MarketplaceAppDTO> entry
keyed by a fixed catalog id; the query service reads it and filters
featured/non-featured in the node process, and registration writes
invalidate it through the provider. Drops the ad-hoc engine:marketplace
cache namespace.
