+++ packages/twenty-server/src/engine/core-entity-cache/types/core-entity-cache-key.type.ts
@@ -1,3 +1,4 @@
+import { type MarketplaceAppDTO } from 'src/engine/core-modules/application/application-marketplace/dtos/marketplace-app.dto';
 import { type FlatUserWorkspace } from 'src/engine/core-modules/user-workspace/types/flat-user-workspace.type';
 import { type FlatUser } from 'src/engine/core-modules/user/types/flat-user.type';
 import { type FlatWorkspace } from 'src/engine/core-modules/workspace/types/flat-workspace.type';
@@ -7,6 +8,7 @@ export type CoreEntityCacheDataMap = {
   user: FlatUser;
   userWorkspaceEntity: FlatUserWorkspace;
   signingKeyPublicKey: string;
+  marketplaceCatalog: Record<string, MarketplaceAppDTO>;
 };
 
 export type CoreEntityCacheKeyName = keyof CoreEntityCacheDataMap;
@@ -16,4 +18,5 @@ export const CORE_ENTITY_CACHE_KEYS: Record<CoreEntityCacheKeyName, string> = {
   user: 'user',
   userWorkspaceEntity: 'user-workspace',
   signingKeyPublicKey: 'signing-key-public-key',
+  marketplaceCatalog: 'marketplace-catalog',
 };
