+++ packages/twenty-server/src/engine/core-modules/application/application-marketplace/constants/marketplace-apps-cache.constant.ts
@@ -1,3 +1,4 @@
-export const MARKETPLACE_APPS_CACHE_TTL_MS = 5 * 60 * 1000;
-
-export const MARKETPLACE_APPS_CACHE_KEY = 'apps:all';
+// Singleton entity id used to cache the whole marketplace catalog as one
+// CoreEntityCache entry. Must be a valid uuid to satisfy CoreEntityCacheService.
+export const MARKETPLACE_CATALOG_CACHE_ENTITY_ID =
+  '00000000-0000-4000-8000-000000000001';
