+++ packages/twenty-server/src/engine/core-modules/application/application-marketplace/marketplace.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 
+import { CoreEntityCacheModule } from 'src/engine/core-entity-cache/core-entity-cache.module';
 import { ApplicationModule } from 'src/engine/core-modules/application/application.module';
 import { ApplicationRegistrationModule } from 'src/engine/core-modules/application/application-registration/application-registration.module';
 import { MarketplaceCatalogSyncCronCommand } from 'src/engine/core-modules/application/application-marketplace/crons/commands/marketplace-catalog-sync.cron.command';
 import { MarketplaceCatalogSyncCronJob } from 'src/engine/core-modules/application/application-marketplace/crons/marketplace-catalog-sync.cron.job';
+import { MarketplaceCatalogCacheProviderService } from 'src/engine/core-modules/application/application-marketplace/marketplace-catalog-cache-provider.service';
 import { MarketplaceCatalogSyncService } from 'src/engine/core-modules/application/application-marketplace/marketplace-catalog-sync.service';
 import { MarketplacePublicResolver } from 'src/engine/core-modules/application/application-marketplace/marketplace-public.resolver';
 import { MarketplaceQueryService } from 'src/engine/core-modules/application/application-marketplace/marketplace-query.service';
@@ -18,6 +20,7 @@ import { MarketplaceCatalogSyncCommand } from 'src/engine/core-modules/applicati
   imports: [
     ApplicationModule,
     ApplicationRegistrationModule,
+    CoreEntityCacheModule,
     FeatureFlagModule,
     PermissionsModule,
     TwentyConfigModule,
@@ -26,6 +29,7 @@ import { MarketplaceCatalogSyncCommand } from 'src/engine/core-modules/applicati
     MarketplaceService,
     MarketplaceCatalogSyncService,
     MarketplaceQueryService,
+    MarketplaceCatalogCacheProviderService,
     MarketplaceCatalogSyncCronJob,
     MarketplaceCatalogSyncCronCommand,
     MarketplaceCatalogSyncCommand,
